refac: move bin related codes from exports to build config
