+++ src/constants.ts
@@ -17,3 +17,9 @@ export const availableESExtensionsRegex = /\.(m|c)?[jt]sx?$/
 export const dtsExtensionRegex = /\.d\.(m|c)?ts$/
 
 export const SRC = 'src'
+
+export const dtsExtentions = {
+  js: '.d.ts',
+  cjs: '.d.cts',
+  mjs: '.d.mts',
+}
